add new BaseCollectionEstimator
