+++ aeon/classification/base.py
@@ -32,22 +32,13 @@ import numpy as np
 import pandas as pd
 from sklearn.utils.multiclass import type_of_target
 
-from aeon.base import BaseEstimator
+from aeon.base import BaseCollectionEstimator
 from aeon.utils.sklearn import is_sklearn_transformer
 from aeon.utils.validation import check_n_jobs
-from aeon.utils.validation._dependencies import _check_estimator_deps
-from aeon.utils.validation.collection import (
-    convert_collection,
-    get_n_cases,
-    has_missing,
-    is_equal_length,
-    is_univariate,
-    resolve_equal_length_inner_type,
-    resolve_unequal_length_inner_type,
-)
+from aeon.utils.validation.collection import get_n_cases
 
 
-class BaseClassifier(BaseEstimator, ABC):
+class BaseClassifier(BaseCollectionEstimator, ABC):
     """
     Abstract base class for time series classifiers.
 
@@ -55,43 +46,31 @@ class BaseClassifier(BaseEstimator, ABC):
 
     Attributes
     ----------
+    fit_time_           : integer, time (in milliseconds) for fit to run.
+    metadata_           : metadata/properties of X seen in fit
+    _n_jobs             : number of threads to use in ``fit`` as determined by
     classes_            : ndarray of class labels, possibly strings
     n_classes_          : integer, number of classes (length of classes_``)
-    fit_time_           : integer, time (in milliseconds) for fit to run.
-    _X_metadata         : metadata/properties of X seen in fit
     _class_dictionary   : dictionary mapping classes_ onto integers
         0...``n_classes_``-1.
-    _n_jobs     : number of threads to use in ``fit`` as determined by
         ``n_jobs``.
     _estimator_type     : string required by sklearn, set to "classifier"
     """
 
     _tags = {
-        "X_inner_mtype": "numpy3D",
-        "capability:multivariate": False,
-        "capability:unequal_length": False,
-        "capability:missing_values": False,
         "capability:train_estimate": False,
-        "capability:contractable": False,
         "capability:multithreading": False,
-        "python_version": None,  # PEP 440 python version specifier to limit versions
+        "capability:contractable": False,
     }
 
     def __init__(self):
         # reserved attributes written to in fit
         self.classes_ = []  # classes seen in y, unique labels
         self.n_classes_ = 0  # number of unique classes in y
-        self.fit_time_ = 0  # time elapsed in last fit call
-        self.metadata_ = {}  # metadata/properties of data seen in fit
         self._class_dictionary = {}
-        self._n_jobs = 1
-
-        # required for compatibility with some sklearn interfaces e.g.
-        # CalibratedClassifierCV
+        # required for compatibility with some sklearn classes
         self._estimator_type = "classifier"
-
         super(BaseClassifier, self).__init__()
-        _check_estimator_deps(self)
 
     def __rmul__(self, other):
         """Magic * method, return concatenated ClassifierPipeline, transformers on left.
@@ -160,18 +139,10 @@ class BaseClassifier(BaseEstimator, ABC):
         # reset estimator at the start of fit
         self.reset()
 
-        # All of this can move up to BaseCollection
+        # All of this can move up to BaseCollection if we enhance fit here with super
         start = int(round(time.time() * 1000))
-        self.metadata_ = self.checkX(X)
+        self.checkX(X)
         X = self.convertX(X)
-        y = self._check_y(y, self.metadata_["n_cases"])
-
-        # escape early and do not fit if only one class label has been seen
-        #   in this case, we later predict the single class label seen
-        if len(self.classes_) == 1:
-            self.fit_time_ = int(round(time.time() * 1000)) - start
-            self._is_fitted = True
-            return self
         multithread = self.get_tag("capability:multithreading")
         if multithread:
             try:
@@ -181,7 +152,15 @@ class BaseClassifier(BaseEstimator, ABC):
                     "self.n_jobs must be set if capability:multithreading is True"
                 )
 
-        # pass coerced and checked data to inner _fit
+        y = self._check_y(y, self.metadata_["n_cases"])
+
+        # escape early and do not fit if only one class label has been seen
+        #   in this case, we later predict the single class label seen
+        if len(self.classes_) == 1:
+            self.fit_time_ = int(round(time.time() * 1000)) - start
+            self._is_fitted = True
+            return self
+        # pass checked and converted data to inner _fit
         self._fit(X, y)
         self.fit_time_ = int(round(time.time() * 1000)) - start
         # this should happen last
@@ -379,120 +358,8 @@ class BaseClassifier(BaseEstimator, ABC):
 
         return dists
 
-    def checkX(self, X):
-        """Check classifier input X is valid.
-
-        Check if the input data is a compatible type, and that this classifier is
-        able to handle the data characteristics. This is done by matching the
-        capabilities of the classifier against the metadata for X for
-        univariate/multivariate, equal length/unequal length and no missing
-        values/missing values.
-
-        Parameters
-        ----------
-        X : object.
-
-        Returns
-        -------
-        bool : True if classifier can deal with X.
-
-        See Also
-        --------
-        convertX : function that converts X after it has been checked.
-
-        Examples
-        --------
-        >>> from aeon.classification.hybrid import HIVECOTEV2
-        >>> import numpy as np
-        >>> X = np.random.random(size=(5,3,10))
-        >>> X[0][1][3] = np.NAN # X is equal length, multivariate, with missing
-        >>> hc = HIVECOTEV2()
-        >>> hc.checkX(X)    # HC2 can handle this
-        True
-        """
-        metadata = _get_metadata(X)
-        # Check classifier capabilities for X
-        allow_multivariate = self.get_tag("capability:multivariate")
-        allow_missing = self.get_tag("capability:missing_values")
-        allow_unequal = self.get_tag("capability:unequal_length")
-
-        # Check capabilities vs input
-        problems = []
-        if metadata["missing_values"] and not allow_missing:
-            problems += ["missing values"]
-        if metadata["multivariate"] and not allow_multivariate:
-            problems += ["multivariate series"]
-        if metadata["unequal_length"] and not allow_unequal:
-            problems += ["unequal length series"]
-
-        if problems:
-            # construct error message
-            problems_and = " and ".join(problems)
-            problems_or = " or ".join(problems)
-            msg = (
-                f"Data seen by instance of {type(self).__name__} has {problems_and}, "
-                f"but {type(self).__name__} cannot handle {problems_or}. "
-            )
-            raise ValueError(msg)
-        return metadata
-
-    def convertX(self, X):
-        """Convert X to type defined by tag X_inner_mtype.
-
-        if self.metadata_ has not been set, it is set here from X, because we need to
-        know if the data is unequal length in order to choose between different
-        allowed input types. If multiple types are allowed by self, then the best
-        one for the data is selected. So, for example, if X_inner_tag is `["np-list",
-        "numpy3D"]` and an `np-list` is passed containing equal length series,
-        X will be converted to numpy3D.
-
-        Parameters
-        ----------
-        X : data structure
-        must be of type aeon.utils.validation.collection.COLLECTIONS_DATA_TYPES.
-
-        Returns
-        -------
-        data structure of type one of self.get_tag("X_inner_mtype").
-
-
-        See Also
-        --------
-        checkX : function that checks X is valid and finds metadata.
-
-        Examples
-        --------
-        >>> from aeon.classification.hybrid import HIVECOTEV2
-        >>> import numpy as np
-        >>> from aeon.utils.validation.collection import get_type
-        >>> X = [np.random.random(size=(5,10), np.random.random(size=(5,10)]
-        >>> get_type(X)
-        np-list
-        >>> hc = HIVECOTEV2()
-        >>> hc.get_tag("X_inner_mtype")
-        ["np-list", "numpy3D"]
-        >>> X = hc.convertX(X)
-        >>> get_type(X)
-        numpy3D
-        """
-        if len(self.metadata_) == 0:
-            metadata = _get_metadata(X)
-        else:
-            metadata = self.metadata_
-        # Convert X to X_inner_mtype if possible
-        inner_type = self.get_tag("X_inner_mtype")
-        if type(inner_type) == list:
-            # If self can handle more than one internal type, resolve correct conversion
-            # If unequal, choose data structure that can hold unequal
-            if metadata["unequal_length"]:
-                inner_type = resolve_unequal_length_inner_type(inner_type)
-            else:
-                inner_type = resolve_equal_length_inner_type(inner_type)
-        X = convert_collection(X, inner_type)
-        return X
-
     def _check_y(self, y, n_cases):
-        # Check y valid input
+        # Check y valid input for classification
         if not isinstance(y, (pd.Series, np.ndarray)):
             raise TypeError(
                 f"y must be a np.array or a pd.Series, but found type: {type(y)}"
@@ -521,14 +388,3 @@ class BaseClassifier(BaseEstimator, ABC):
         for index, class_val in enumerate(self.classes_):
             self._class_dictionary[class_val] = index
         return y
-
-
-def _get_metadata(X):
-    # Get and store X meta data.
-    metadata = {}
-    metadata["multivariate"] = not is_univariate(X)
-    metadata["missing_values"] = has_missing(X)
-    metadata["unequal_length"] = not is_equal_length(X)
-    metadata["n_cases"] = get_n_cases(X)
-
-    return metadata
